# docs/swagger.yaml
basePath: /
definitions:
  models.ComposeActionRequest:
    properties:
      name:
        example: my-app
        type: string
    type: object
  models.ComposeContainerInfo:
    properties:
      id:
        example: a1b2c3d4e5f6
        type: string
      image:
        example: nginx:latest
        type: string
      name:
        example: /app1-web
        type: string
      ports:
        example: 8080:80 443:443
        type: string
      status:
        example: Up 3 minutes
        type: string
    type: object
  models.ComposeStatusResponse:
    properties:
      apps:
        additionalProperties:
          items:
            $ref: '#/definitions/models.ComposeContainerInfo'
          type: array
        type: object
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal Server Error
        type: string
    type: object
...
  models.ListComposeResponse:
    properties:
      apps:
        example:
        - '["app1"'
        - ' "app2"]'
        items:
          type: string
        type: array
    type: object
...
  models.RunAnsibleRequest:
    properties:
      inventory:
        example: /etc/ansible/hosts
        type: string
      playbook:
        example: site.yml
        type: string
    type: object
  models.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Container stopped successfully
        type: string
    type: object
host: api.xkkk.online
info:
  contact:
    email: kliu4403@gmail.com
    name: Dev Team
    url: https://www.xkkk.online
  description: 自动化部署平台接口文档，支持容器、Compose、文件、Ansible管理。
  title: Auto Deploy Platform API
  version: "1.0"
paths:
  /ansible/run:
    post:
      consumes:
      - application/json
      description: 通过提供 Inventory 文件和 Playbook 路径，执行 Ansible 并实时返回日志流
      parameters:
      - description: Ansible 执行参数
        in: body
        name: ansible
        required: true
        schema:
          $ref: '#/definitions/models.RunAnsibleRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: 执行日志流（stream）
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ansible 执行失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 执行 Ansible Playbook
      tags:
      - 自动化部署
  /compose/delete:
    post:
      consumes:
      - application/json
      description: 删除指定 Compose 应用及其目录
      parameters:
      - description: Compose 应用名称
        in: body
        name: compose
        required: true
        schema:
          $ref: '#/definitions/models.ComposeActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 删除 Compose 应用
      tags:
      - Compose管理
  /compose/list:
    get:
      description: 列出当前存在的所有 Compose 应用
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回 Compose 应用列表
          schema:
            $ref: '#/definitions/models.ListComposeResponse'
        "500":
          description: 读取目录失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 获取 Compose 应用列表
      tags:
      - Compose管理
  /compose/start:
    post:
      consumes:
      - application/json
      description: 通过应用名称启动对应 Compose 应用
      parameters:
      - description: Compose 应用名称
        in: body
        name: compose
        required: true
        schema:
          $ref: '#/definitions/models.ComposeActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 启动成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 启动失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 启动 Compose 应用
      tags:
      - Compose管理
  /compose/status:
    get:
      description: 查看各 Compose 应用包含的容器及运行状态
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回 Compose 容器状态
          schema:
            $ref: '#/definitions/models.ComposeStatusResponse'
        "500":
          description: Docker client 初始化或容器列表失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 获取 Compose 应用状态
      tags:
      - Compose管理
  /compose/stop:
    post:
      consumes:
      - application/json
      description: 通过应用名称停止对应 Compose 应用
      parameters:
      - description: Compose 应用名称
        in: body
        name: compose
        required: true
        schema:
          $ref: '#/definitions/models.ComposeActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 停止成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 停止失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 停止 Compose 应用
      tags:
      - Compose管理
  /compose/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传并保存 Docker Compose 文件
      parameters:
      - description: Compose 文件名称
        in: formData
        name: name
        required: true
        type: string
      - description: Compose 文件 (YAML格式)
        in: formData
        name: compose_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 上传 Compose 文件
      tags:
      - Compose管理
...
schemes:
- https
securityDefinitions:
  BearerToken:
    description: '请输入 Bearer Token 认证，比如: Bearer eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"